Rename lib back to core
