+++ src/page_01/problem_18.py
@@ -1,4 +1,4 @@
-from lib import gridutils
+from core import gridutils
 
 TRIANGLE = """
 75
